Update idc/atmel-touchscreen.idc
It seems that nobody noticed this, but the touch screen configuration is wrong.
That leads to inablity to detect touch size and general touchscreen 'precision' is bad in some situations.
This file fixes the touch size reporting (which is actually supported by our phone).
More on touch configuration: http://source.android.com/tech/input/touch-devices.html